the location of the components has been changed, now UserItem and UserRoleItem are separate components
