+++ src/entities/User/ui/UserDetails.module.css
@@ -1,16 +1,6 @@
 .wrapper {
     display: flex;
 }
-.userSmallInfo {
-    flex: 1;
-    cursor: pointer;
-    margin: 2px;
-    padding: 5px 4px;
-    border: 1px solid #d3d3d3;
-    border-radius: 4px;
-    color: #3a6565;
-    background-color: #f2ffff;
-}
 .userFullInfo {
     flex: 1;
     margin: 2px;
@@ -20,11 +10,6 @@
     color: #3a6565;
     background-color: #f2ffff;
 }
-.userSmallInfo:hover {
-    background-color: #3a6565;
-    color: white;
-}
-
 .userTextInfo {
     display: flex;
 }
@@ -32,35 +17,25 @@
     display: flex;
     flex-direction: column;
 }
-
+.inputLogin {
+    padding: 4px;
+}
+.inputEmail {
+    padding: 4px;
+}
 .userDate {
     margin-left: 40px;
 }
-.test2 {
-    font-weight:bold;
-}
-
 .roleGroup {
     display: flex;
     margin-top: 10px;
     align-items: center;
 }
-.role {
-    border: 1px solid black;
-    margin-left: 5px;
-    padding: 2px;
-    margin-top: 5px;
-}
+
 .buttonAddRole{
     margin-left: 10px;
     margin-top: 5px;
 }
 .modal {
     padding: 10px;
-}
-.inputLogin {
-    padding: 4px;
-}
-.inputEmail {
-    padding: 4px;
 }
