+++ src/features/UserItem/index.js
@@ -1,0 +1,1 @@
+export * from './ui/UserItem'
