+++ src/features/UserRoleItem/ui/UserRoleItem.js
@@ -1,7 +1,6 @@
-import style from "./User.module.css";
+import style from './UserRoleItem.module.css';
 import React from "react";
-
-export const UserRoles = ({role}) => {
+export const UserRoleItem = ({role}) => {
     return <div className={style.wrapper}>
         <div className={style.role}>{role}</div>
     </div>
