+++ src/features/UserRoleItem/ui/UserRoleItem.module.css
@@ -1,0 +1,9 @@
+.wrapper {
+    display: flex;
+}
+.role {
+    border: 1px solid black;
+    margin-left: 5px;
+    padding: 2px;
+    margin-top: 5px;
+}
